Upload Pipeline v2

+++ wcfsetup/install/files/lib/data/attachment/AttachmentEditor.class.php
@@ -64,6 +64,10 @@ class AttachmentEditor extends DatabaseObjectEditor
      */
     public function deleteFiles()
     {
+        if ($this->fileID !== null) {
+            return;
+        }
+
         @\unlink($this->getLocation());
         if ($this->tinyThumbnailType) {
             @\unlink($this->getTinyThumbnailLocation());
